[Fix #3411] Avoid crash in Performance::CaseWhenSplat#autocorrect
A special case is when there is only one `when` branch, and it has
multiple conditions, with a splat followed by non-splat. This change
only fixes the crash. I think it's possible that reordering the
conditions on the line leads to faster code, but that should be checked.
It's also a more complicated change that could lead to new bugs.