Preparing for 1.1.0 release
Several performance cops were moved to the style department in the
RuboCop 0.67.
https://github.com/rubocop-hq/rubocop/releases/tag/v0.67.0

According to that, the next RuboCop Performance release requires
RuboCop 0.67 or higher.

The next RuboCop Perfomance release removes some perfomance cops and
bumps the dependent RuboCop version. So this PR makes the next release
version to 1.1.0 instead of 1.0.1.